requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
